test(celx): add opt-in benchmarks for removeKeys/filterKeys
